backend/src/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Counter==1.0.0
cryptography==3.4.7
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
drf-social-oauth2==1.0.8
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
Pillow==8.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
